Better span labels
