+++ core/deployment/src/main/java/io/quarkus/deployment/builditem/FeatureBuildItem.java
@@ -16,6 +16,7 @@ public final class FeatureBuildItem extends MultiBuildItem {
     public static final String CDI = "cdi";
     public static final String DYNAMODB = "dynamodb";
     public static final String ELASTICSEARCH_REST_CLIENT = "elasticsearch-rest-client";
+    public static final String GCP_STACKDRIVER = "gcp-stackdriver";
     public static final String FLYWAY = "flyway";
     public static final String HIBERNATE_ORM = "hibernate-orm";
     public static final String HIBERNATE_ORM_PANACHE = "hibernate-orm-panache";
